+++ core/designsystem/src/main/java/com/saegil/designsystem/component/SaegilLoadingWheel.kt
@@ -1,0 +1,19 @@
+package com.saegil.designsystem.component
+
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SaegilLoadingWheel(
+    modifier: Modifier = Modifier
+) {
+    CircularProgressIndicator(
+        modifier = modifier.width(48.dp),//부모로부터 받은 modifier를 중앙배치용으로 사용
+        color = MaterialTheme.colorScheme.secondary,
+        trackColor = MaterialTheme.colorScheme.surfaceVariant,
+    )
+}
